tests(Warnings): update tests
